python-server/requirements.txt
flask==3.1.0
flask-cors==5.0.0
ytmusicapi==1.8.1
python-dotenv==1.0.1
gunicorn==23.0.0
